refactor: add ModalSwitchComponent
